Move method to TestFunction
